+++ MacPacker/Core/UrlHandling/AppUrlExtractToFolderHandler.swift
@@ -15,7 +15,7 @@ class AppUrlExtractToFolderHandler: AppUrlHandler {
             requestAccessToDir(for: appUrl.target) { response, url in
                 if response == .OK {
                     if let url,
-                       let archiveHandler = ArchiveHandlerRegistry.shared.handler(for: fileUrl)
+                       let archiveHandler = ArchiveTypeRegistry.shared.handler(for: fileUrl)
                     {
                         let folderName = fileUrl.deletingPathExtension().lastPathComponent
                         let folderUrl = url.appendingPathComponent(folderName)